Полностью разделена игра на несколько файлов
